+++ lib/src/chip/chip.dart
@@ -1,0 +1,3 @@
+export 'sbb_chip.dart';
+export 'theme/sbb_chip_style.dart';
+export 'theme/sbb_chip_theme_data.dart';